requirements.txt
et-xmlfile==1.1.0
openpyxl==3.0.9
Pillow==8.4.0
